recent changes

+++ src/TELEGRAM_API.java
@@ -2,6 +2,7 @@ import Constants.Constants;
 import InstagramItems.InstagramPost;
 import InstagramItems.InstagramReel;
 import InstagramItems.InstagramStory;
+import Singletons.Instagram;
 import Singletons.Telegram;
 import com.pengrad.telegrambot.TelegramBot;
 import com.pengrad.telegrambot.UpdatesListener;
@@ -20,6 +21,7 @@ public class TELEGRAM_API {
         bot.setUpdatesListener(updates -> {
             for (Update update : updates) {
                 if (update.message() == null) continue;
+                if (update.message().text() == null) continue;
                 String answer = "";
                 long chatId = update.message().chat().id();
                 boolean is_admin = false;
@@ -49,7 +51,25 @@ public class TELEGRAM_API {
             scheduleShutDown();
             return "Приложение отключится через минуту";
         }
-        if (message.equals("/story")) {
+
+        // Message:
+        // Login
+        // Username: newUserName
+        // Password: newPassword
+        if (message.toLowerCase().contains("login")) {
+            String username = "";
+            String password = "";
+            String[] msg = message.split("\n");
+            for (String m : msg) {
+                if (m.toLowerCase().contains("username:")) {
+                    username = m.split(":")[1].trim();
+                } else if (m.toLowerCase().contains("password:")) {
+                    password = m.split(":")[1].trim();
+                }
+            }
+            Instagram.reLogin(username, password);
+            return "Выполнен повторный вход в аккаунт";
+        } else if (message.equals("/story")) {
             try {
                 Updater.storyUpdater();
                 return "Сториз были проверены";
@@ -58,8 +78,7 @@ public class TELEGRAM_API {
                 e.printStackTrace();
                 return "В настоящий момент проверку сториз выполнить невозможно.";
             }
-        }
-        if (message.equals("/post")) {
+        } else if (message.equals("/post")) {
             try {
                 Updater.postUpdater();
                 return "Посты были проверены";
@@ -68,13 +87,11 @@ public class TELEGRAM_API {
                 e.printStackTrace();
                 return "В настоящий момент проверку постов выполнить невозможно.";
             }
-        }
-        if (Pattern.matches("[0-9][0-9][0-9][0-9][0-9][0-9]", message)) {
+        } else if (Pattern.matches("[0-9][0-9][0-9][0-9][0-9][0-9]", message)) {
             Constants.sixDigits = message;
             Constants.hasDigits = true;
             return "Шестизначный код обработан";
-        }
-        if (message.contains("instagram.com/") || message.contains("youtu.be") || message.contains("youtube.com")) {
+        } else if (message.contains("instagram.com/") || message.contains("youtu.be") || message.contains("youtube.com")) {
             String answer = TELEGRAM_API.responseTelegram(message);
             if (answer != Constants.successStory && answer != Constants.successPost
                     && answer != Constants.successReel && answer != Constants.successYoutube) {
@@ -115,6 +132,17 @@ public class TELEGRAM_API {
 
     private static String responseTelegram(String message) {
         String answer = "";
+        String flag = "";
+        String possibleFlag = message.split("-")[message.split("-").length - 1].toLowerCase();
+        if (possibleFlag.equals(("v"))) {
+            flag = Constants.flagVk;
+            System.out.println("Posting to VK only");
+        }
+        else if (possibleFlag.equals("t")) {
+            flag = Constants.flagTg;
+            System.out.println("Posting to TELEGRAM only");
+        }
+
         if (message.contains("youtu.be") || message.contains("youtube.com")) {
             String[] msg = message.split("\n");
             String youtubeUrl = "";
@@ -126,7 +154,7 @@ public class TELEGRAM_API {
                     if (!m.isBlank())
                         messagePost = m.trim();
             }
-            answer = postingVideoFromYoutube(youtubeUrl, messagePost);
+            answer = postingVideoFromYoutube(youtubeUrl, messagePost, flag);
         } else if (message.length() < 30)
             answer = "Ты отправил что-то не то. Попробуй еще раз.";
         else {
@@ -143,7 +171,7 @@ public class TELEGRAM_API {
                             story.setLink(links[i]);
                             storiesToPost[i] = story;
                         }
-                        answer = postingStoriesNotification(storiesToPost);
+                        answer = postingStoriesNotification(storiesToPost, flag);
                     } catch (Exception e) {
                         e.printStackTrace();
                     }
@@ -166,9 +194,9 @@ public class TELEGRAM_API {
                                 return "Неверно указан индекс!";
                             }
                         }
-                        answer = postingPostNotification(post, indexes);
+                        answer = postingPostNotification(post, indexes, flag);
                     } else {
-                        answer = postingPostNotification(post);
+                        answer = postingPostNotification(post, flag);
                     }
                 } catch (Exception e) {
                     e.printStackTrace();
@@ -178,7 +206,7 @@ public class TELEGRAM_API {
                 try {
                     InstagramReel reel = INST_API.getUserReelByLink(message);
                     reel.setLink(message);
-                    answer = postingReelNotification(reel);
+                    answer = postingReelNotification(reel, flag);
                 } catch (Exception e) {
                     e.printStackTrace();
                 }
@@ -191,60 +219,71 @@ public class TELEGRAM_API {
         return answer;
     }
 
-    public static String postingVideoFromYoutube(String youtubeUrl, String message) {
+    public static String postingVideoFromYoutube(String youtubeUrl, String message, String flag) {
         StringJoiner joiner = new StringJoiner("\n\n");
         boolean successVK = false;
         boolean successTG = false;
 
-        try {
-            VK_API.postVideoFromYouTube(youtubeUrl, message);
-            successVK = true;
-        } catch (Exception e) {
-            joiner.add("Не удалось опубликовать видео в вк\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
+        if (flag.isEmpty() || flag == Constants.flagVk) {
+            try {
+                VK_API.postVideoFromYouTube(youtubeUrl, message);
+                successVK = true;
+            } catch (Exception e) {
+                joiner.add("Не удалось опубликовать видео в вк\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
+            }
         }
 
-        try {
-            ChannelPosting.sendVideoLink(youtubeUrl, message);
-            successTG = true;
-        } catch (Exception e) {
-            joiner.add("Не удалось опубликовать пост в тг\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
+        if (flag.isEmpty() || flag == Constants.flagTg) {
+            try {
+                ChannelPosting.sendVideoLink(youtubeUrl, message);
+                successTG = true;
+            } catch (Exception e) {
+                joiner.add("Не удалось опубликовать пост в тг\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
+            }
         }
 
-        if (successTG && successVK) return Constants.successYoutube;
+        if ((flag.isEmpty() && successTG && successVK) ||
+                (!flag.isEmpty() && (successTG || successVK))) return Constants.successYoutube;
         return joiner.toString();
     }
 
-    public static String postingPostNotification (InstagramPost post)  {
+    public static String postingPostNotification (InstagramPost post, String flag)  {
         if (post.getUser() == null) return "Null Pointer Exception >_<" + Constants.getStackTrace();
 
         StringJoiner joiner = new StringJoiner("\n\n");
         boolean successVK = false;
         boolean successTG = false;
         String vkLink = null;
-        try {
-            vkLink = VK_API.postPost(post);
-            successVK = true;
-        } catch (Exception e) {
-            joiner.add("Не удалось опубликовать пост в вк\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
+
+        if (flag.isEmpty() || flag == Constants.flagVk) {
+            try {
+                vkLink = VK_API.postPost(post);
+                successVK = true;
+            } catch (Exception e) {
+                joiner.add("Не удалось опубликовать пост в вк\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
+            }
         }
 
-        try {
-            ChannelPosting.sendMessageToChannel(post);
-            successTG = true;
-        } catch (Exception e) {
-            joiner.add("Не удалось опубликовать пост в тг\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
-            if (vkLink != null)
-                ChannelPosting.sendLinks(vkLink, post.getLink());
-        } finally {
-            for (File f : post.getMedia())
-                f.delete();
+        if (flag.isEmpty() || flag == Constants.flagTg) {
+            try {
+                ChannelPosting.sendMessageToChannel(post);
+                successTG = true;
+            } catch (Exception e) {
+                joiner.add("Не удалось опубликовать пост в тг\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
+                if (vkLink != null)
+                    ChannelPosting.sendLinks(vkLink, post.getLink());
+            } finally {
+                for (File f : post.getMedia())
+                    f.delete();
+            }
         }
 
-        if (successTG && successVK) return Constants.successPost;
+        if ((flag.isEmpty() && successTG && successVK) ||
+                (!flag.isEmpty() && (successTG || successVK))) return Constants.successPost;
         return joiner.toString();
     }
 
-    public static String postingPostNotification (InstagramPost post, Set<Integer> indexes)  {
+    public static String postingPostNotification (InstagramPost post, Set<Integer> indexes, String flag)  {
         if (post.getUser() == null) return "Null Pointer Exception >_<" + Constants.getStackTrace();
 
         for (int index : indexes)
@@ -255,30 +294,36 @@ public class TELEGRAM_API {
         boolean successVK = false;
         boolean successTG = false;
         String vkLink = null;
-        try {
-            vkLink = VK_API.postPost(post, indexes);
-            successVK = true;
-        } catch (Exception e) {
-            joiner.add("Не удалось опубликовать пост в вк\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
-            if (vkLink != null)
-                ChannelPosting.sendLinks(vkLink, post.getLink());
+
+        if (flag.isEmpty() || flag == Constants.flagVk) {
+            try {
+                vkLink = VK_API.postPost(post, indexes);
+                successVK = true;
+            } catch (Exception e) {
+                joiner.add("Не удалось опубликовать пост в вк\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
+                if (vkLink != null)
+                    ChannelPosting.sendLinks(vkLink, post.getLink());
+            }
         }
 
-        try {
-            ChannelPosting.sendMessageToChannel(post, indexes);
-            successTG = true;
-        } catch (Exception e) {
-            joiner.add("Не удалось опубликовать пост в тг\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
-        } finally {
-            for (File f : post.getMedia())
-                f.delete();
+        if (flag.isEmpty() || flag == Constants.flagTg) {
+            try {
+                ChannelPosting.sendMessageToChannel(post, indexes);
+                successTG = true;
+            } catch (Exception e) {
+                joiner.add("Не удалось опубликовать пост в тг\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
+            } finally {
+                for (File f : post.getMedia())
+                    f.delete();
+            }
         }
 
-        if (successTG && successVK) return Constants.successPost;
+        if ((flag.isEmpty() && successTG && successVK) ||
+                (!flag.isEmpty() && (successTG || successVK))) return Constants.successPost;
         return joiner.toString();
     }
 
-    public static String postingStoriesNotification (InstagramStory[] stories) {
+    public static String postingStoriesNotification (InstagramStory[] stories, String flag) {
         for (InstagramStory story : stories) {
             if (story.getUser() == null)
                 return "Null Pointer Exception >_<" + Constants.getStackTrace();
@@ -287,51 +332,63 @@ public class TELEGRAM_API {
         StringJoiner joiner = new StringJoiner("\n\n");
         boolean successVK = false;
         boolean successTG = false;
-        try {
-            VK_API.postStory(stories);
-            successVK = true;
-        } catch (Exception e) {
-            joiner.add("Не удалось опубликовать стори в вк\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
+
+        if (flag.isEmpty() || flag == Constants.flagVk) {
+            try {
+                VK_API.postStory(stories);
+                successVK = true;
+            } catch (Exception e) {
+                joiner.add("Не удалось опубликовать стори в вк\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
+            }
         }
 
-        try {
-            ChannelPosting.sendMessageToChannel(stories);
-            successTG = true;
-        } catch (Exception e) {
-            joiner.add("Не удалось опубликовать стори в тг\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
-        } finally {
-            for (InstagramStory story : stories)
-                story.getMedia().delete();
+        if (flag.isEmpty() || flag == Constants.flagTg) {
+            try {
+                ChannelPosting.sendMessageToChannel(stories);
+                successTG = true;
+            } catch (Exception e) {
+                joiner.add("Не удалось опубликовать стори в тг\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
+            } finally {
+                for (InstagramStory story : stories)
+                    story.getMedia().delete();
+            }
         }
 
-        if (successTG && successVK) return Constants.successStory;
+        if ((flag.isEmpty() && successTG && successVK) ||
+                (!flag.isEmpty() && (successTG || successVK))) return Constants.successStory;
         return joiner.toString();
     }
 
-    public static String postingReelNotification (InstagramReel reel) {
+    public static String postingReelNotification (InstagramReel reel, String flag) {
         if (reel.getUser() == null)
             return "Null Pointer Exception >_<" + Constants.getStackTrace();
 
         StringJoiner joiner = new StringJoiner("\n\n");
         boolean successVK = false;
         boolean successTG = false;
-        try {
-            VK_API.postReel(reel);
-            successVK = true;
-        } catch (Exception e) {
-            joiner.add("Не удалось опубликовать рил в вк\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
+
+        if (flag.isEmpty() || flag == Constants.flagVk) {
+            try {
+                VK_API.postReel(reel);
+                successVK = true;
+            } catch (Exception e) {
+                joiner.add("Не удалось опубликовать рил в вк\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
+            }
         }
 
-        try {
-            ChannelPosting.sendMessageToChannel(reel);
-            successTG = true;
-        } catch (Exception e) {
-            joiner.add("Не удалось опубликовать рил в тг\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
-        } finally {
-            reel.getVideo().delete();
+        if (flag.isEmpty() || flag == Constants.flagTg) {
+            try {
+                ChannelPosting.sendMessageToChannel(reel);
+                successTG = true;
+            } catch (Exception e) {
+                joiner.add("Не удалось опубликовать рил в тг\n\n" + e.getMessage() + "\n\n" + Constants.getStackTrace());
+            } finally {
+                reel.getVideo().delete();
+            }
         }
 
-        if (successTG && successVK) return Constants.successReel;
+        if ((flag.isEmpty() && successTG && successVK) ||
+                (!flag.isEmpty() && (successTG || successVK))) return Constants.successReel;
         return joiner.toString();
     }
 }
